refactor: Add file download functionality to ResourceVersionList component
- Enhanced previously functional-less buttons by incorporating a download function. This download functionality modularizes the logic initially used in the ControlPanel component into a custom hook, enabling reuse across both components.
